requirements.txt
loguru==0.7.2
maafw==5.2.6